refactor: move ssl-content and dump-responses to network
