+++ cyberdrop_dl/clients/http.py
@@ -92,14 +92,14 @@ class HTTPClient:
     _download_session: aiohttp.ClientSession = dataclasses.field(init=False)
 
     def __post_init__(self) -> None:
-        self._ssl_context = tcp.create_ssl_context(self.config.ssl_context)
+        self._ssl_context = tcp.create_ssl_context(self.config.network.ssl_context)
         self.global_rate_limiter = aio.RateLimiter.w_no_burst(self.config.network.rate_limit)
         self.global_download_limiter = asyncio.Semaphore(self.config.downloads.concurrency)
 
     @staticmethod
     def from_manager(manager: Manager) -> HTTPClient:
         client = HTTPClient(config=manager.config, impersonate=manager.cli_args.impersonate)
-        if manager.config.dump_responses:
+        if manager.config.network.dump_responses:
             client.request_done_callback = manager.logs.write_response
 
         return client
@@ -360,7 +360,7 @@ def _create_curl_session(config: Config) -> AsyncSession[CurlResponse]:
         loop=loop,
         async_curl=acurl,
         impersonate="chrome",
-        verify=bool(config.ssl_context),
+        verify=bool(config.network.ssl_context),
         proxy=str(proxy) if (proxy := config.network.proxy) else None,
         timeout=config.network.curl_timeout,
         max_redirects=8,
